Cleaned up code to be more readable
