+++ classes/deck_class.py
@@ -1,15 +1,16 @@
 from classes.card_class import Card
 
+
 class Deck:
 
     def __init__(self):
         self.deck = []
 
     def generate(self):
-        card_types = ["Hearts", "Diamonds", "Spades", "Clubs"]
+        card_types = ["♥", "🔹", "♠", "♣"]
 
-        for type in card_types:
-            for numbers in range(1, 13): # 4 different types and 12 numbers for each type
-                self.deck.append(Card(type, numbers))
+        for card_type in card_types:
+            for numbers in range(1, 13):  # 4 different types and 12 numbers for each type
+                self.deck.append(Card(card_type, numbers))
 
         return self.deck
